# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/numbers/{value}": {
            "get": {
                "description": "Searches for the index for the given value",
                "summary": "Search Index For Value Endpoint",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Value to look for.",
                        "name": "value",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/services.SearchResultDto"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/entrypoints.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/entrypoints.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "entrypoints.ErrorResponse": {
            "type": "object",
            "properties": {
                "details": {
                    "type": "string"
                }
            }
        },
        "services.SearchResultDto": {
            "type": "object",
            "properties": {
                "found": {
                    "type": "boolean"
                },
                "index": {
                    "type": "integer"
                },
                "value": {
                    "type": "integer"
                }
            }
        }
    }
}